fixed paths related to the move
